Rename const to constants; remove needless extension
